Form serialization fixes and improvements
AVIF/AVSK is no longer saved for non-skill actor values.

BGSIdleCollection data is no longer saved if the collection has no idles and the data is zero.

Actor values now have hardcoded form data set up properly.

The hardcoded, legacy "reanimate eyes" form now has its full name set properly.

Fixed loading and saving of FURN/WBDT skill indices.

Fixed width of LTEX/INAM flags.

PERK entry points now skip saving empty condition groups, consistent with the CK.

+++ native/src/DovahKit/dovah/forms/LandTexture.h
@@ -15,7 +15,7 @@ namespace dovah::loaded_forms {
 
          struct remaster_flag {
             remaster_flag() = delete;
-            enum type : uint8_t {
+            enum type : uint32_t {
                is_snow = 0x01,
             };
          };
